third_party/WebKit/LayoutTests/http/tests/inspector/service-workers/service-worker-agents-expected.txt

Tests the way service workers don't enable DOM agent and does enable Debugger agent.

Debugger-related command should be issued: {
    "id": "<id>",
    "method": "Target.sendMessageToTarget",
    "params": {
        "targetId": "<id>",
        "message": {
            "id": "<id>",
            "method": "Target.sendMessageToTarget",
            "params": {
                "targetId": "<id>",
                "message": {
                    "id": "<id>",
                    "method": "Debugger.enable"
                }
            }
        }
    }
}
Suspending targets.
Resuming targets.
Debugger-related command should be issued: {
    "id": "<id>",
    "method": "Target.sendMessageToTarget",
    "params": {
        "targetId": "<id>",
        "message": {
            "id": "<id>",
            "method": "Target.sendMessageToTarget",
            "params": {
                "targetId": "<id>",
                "message": {
                    "id": "<id>",
                    "method": "Debugger.enable"
                }
            }
        }
    }
}
